JVNGRS-45: changed totalRevenue type from BigDecimal to Double

+++ back-end/services/analytics-service/src/main/java/com/project/mentorship/service/analytics/api/dto/StatisticsDto.java
@@ -1,9 +1,8 @@
 package com.project.mentorship.service.analytics.api.dto;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
-public record StatisticsDto(UUID id, OffsetDateTime date, Integer totalReservations, BigDecimal totalRevenue,
+public record StatisticsDto(UUID id, OffsetDateTime date, Integer totalReservations, Double totalRevenue,
 		OffsetDateTime createdAt) {
 }
