+++ back-end/services/analytics-service/src/test/java/com/project/mentorship/service/analytics/persistance/StatisticsRepositoryTest.java
@@ -3,7 +3,6 @@ package com.project.mentorship.service.analytics.persistance;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import com.project.mentorship.service.analytics.domain.Statistics;
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.UUID;
@@ -16,7 +15,7 @@ class StatisticsRepositoryTest {
 		// Given
 		StatisticsRepository statisticsRepository = new StatisticsRepository();
 		Statistics statistics = new Statistics(UUID.randomUUID(),
-				OffsetDateTime.of(2025, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC), 10, BigDecimal.valueOf(250),
+				OffsetDateTime.of(2025, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC), 10, Double.valueOf(250.0),
 				OffsetDateTime.of(2025, 1, 1, 1, 0, 0, 0, ZoneOffset.UTC));
 
 		// When
